[web] Refactoring web common utils
Move web specific util functions to `web/shared/util.py`.
