+++ analyzer/codechecker_analyzer/analyzers/clangsa/statistics_collector.py
@@ -21,8 +21,7 @@ import re
 
 from codechecker_common.logger import get_logger
 
-from codechecker_analyzer.analyzer_env import \
-    extend_analyzer_cmd_with_resource_dir
+from codechecker_analyzer.env import extend_analyzer_cmd_with_resource_dir
 
 LOG = get_logger('analyzer')
