+++ codechecker_common/util.py
@@ -11,19 +11,11 @@ from __future__ import division
 from __future__ import absolute_import
 
 import argparse
-import datetime
-import hashlib
 import io
 import json
 import os
 import re
-import shutil
-import socket
-import stat
-import subprocess
-import tempfile
 import time
-import uuid
 
 from threading import Timer
 
@@ -35,35 +27,6 @@ from codechecker_common.logger import get_logger
 LOG = get_logger('system')
 
 
-class CmdLineOutputEncoder(json.JSONEncoder):
-    def default(self, o):
-        d = {}
-        d.update(o.__dict__)
-        return d
-
-
-class DBSession(object):
-    """
-    Requires a session maker object and creates one session which can be used
-    in the context.
-
-    The session will be automatically closed, but commiting must be done
-    inside the context.
-    """
-    def __init__(self, session_maker):
-        self.__session = None
-        self.__session_maker = session_maker
-
-    def __enter__(self):
-        # create new session
-        self.__session = self.__session_maker()
-        return self.__session
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if self.__session:
-            self.__session.close()
-
-
 class RawDescriptionDefaultHelpFormatter(
         argparse.RawDescriptionHelpFormatter,
         argparse.ArgumentDefaultsHelpFormatter
@@ -73,56 +36,6 @@ class RawDescriptionDefaultHelpFormatter(
     descriptions.
     """
     pass
-
-
-def get_free_port():
-    """ Get a free port from the OS. """
-
-    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    s.bind(('', 0))
-    free_port = s.getsockname()[1]
-    s.close()
-
-    return free_port
-
-
-def is_localhost(address):
-    """
-    Check if address is one of the valid values and try to get the
-    IP-addresses from the system.
-    """
-
-    valid_values = ['localhost', '0.0.0.0', '*', '::1']
-
-    try:
-        valid_values.append(socket.gethostbyname('localhost'))
-    except socket.herror:
-        LOG.debug("Failed to get IP address for localhost.")
-
-    try:
-        valid_values.append(socket.gethostbyname(socket.gethostname()))
-    except (socket.herror, socket.gaierror):
-        LOG.debug("Failed to get IP address for hostname '%s'",
-                  socket.gethostname())
-
-    return address in valid_values
-
-
-def get_tmp_dir_hash():
-    """Generate a hash based on the current time and process id."""
-
-    pid = os.getpid()
-    time = datetime.datetime.now()
-
-    data = str(pid) + str(time)
-
-    dir_hash = hashlib.md5()
-    dir_hash.update(data)
-
-    LOG.debug('The generated temporary directory hash is %s.',
-              dir_hash.hexdigest())
-
-    return dir_hash.hexdigest()
 
 
 def find_by_regex_in_envpath(pattern, environment):
@@ -178,28 +91,6 @@ def get_binary_in_path(basename_list, versioning_pattern, env):
 
         # Return the one earliest in PATH.
         return files[0]
-
-
-def call_command(command, env=None, cwd=None):
-    """ Call an external command and return with (output, return_code)."""
-
-    try:
-        LOG.debug('Run %s', ' '.join(command))
-        out = subprocess.check_output(command,
-                                      bufsize=-1,
-                                      env=env,
-                                      stderr=subprocess.STDOUT,
-                                      cwd=cwd)
-        LOG.debug(out)
-        return out, 0
-    except subprocess.CalledProcessError as ex:
-        LOG.debug('Running command "%s" Failed.', ' '.join(command))
-        LOG.debug(str(ex.returncode))
-        LOG.debug(ex.output)
-        return ex.output, ex.returncode
-    except OSError as oerr:
-        LOG.warning(oerr.strerror)
-        return oerr.strerror, oerr.errno
 
 
 def kill_process_tree(parent_pid, recursive=False):
@@ -307,163 +198,6 @@ def setup_process_timeout(proc, timeout,
     return __cleanup_timeout
 
 
-def get_default_workspace():
-    """
-    Default workspace in the users home directory.
-    """
-    workspace = os.path.join(os.path.expanduser("~"), '.codechecker')
-    return workspace
-
-
-def __strip_protocol_from_url(url):
-    """
-    Strip the PROTOCOL specifier from an URL string and return it, along with
-    the remainder of the URL.
-    """
-    if '://' not in url:
-        return None, url
-
-    parts = url.split('://', 1)
-    protocol = parts[0]
-    url = url.replace(parts[0] + '://', '').lstrip('/').rstrip('/')
-
-    return protocol, url
-
-
-def expand_whole_protocol_and_port(protocol=None, port=None):
-    """
-    Calculate a full protocol and port value from the possibly None protocol
-    and port values given. This method helps to default port numbers to
-    connection protocols understood by CodeChecker.
-    """
-
-    if protocol:
-        if protocol == 'http':
-            portnum = 80
-        elif protocol == 'https':
-            portnum = 443
-        else:
-            raise ValueError("'{0}' is not a protocol understood by "
-                             "CodeChecker".format(protocol))
-    else:
-        protocol = 'http'
-        portnum = 8001
-
-    if port:
-        portnum = port
-
-    return protocol, portnum
-
-
-def split_server_url(url):
-    """
-    Splits the given CodeChecker server URL into its parts.
-
-    The format of a valid URL is:
-      protocol://host:port/
-    where
-      * Protocol: HTTP or HTTPS
-      * Host: The server's host name or IP address
-      * Port: The server's port number
-
-    As a shortcut, the following formats are also valid:
-      hostname           (means: http://hostname:8001)
-    """
-
-    LOG.debug("Parsing server url '%s'", url)
-    protocol, url = __strip_protocol_from_url(url)
-
-    # A valid product_url looks like this: 'http://localhost:8001/Product'.
-    protocol, port = expand_whole_protocol_and_port(protocol, None)
-    host = 'localhost'
-    try:
-        parts = url.split('/', 1)
-
-        # Something is either a hostname, or a host:port.
-        server_addr = parts[0].split(":")
-        if len(server_addr) == 2:
-            host, port = server_addr[0], int(server_addr[1])
-        elif len(server_addr) == 1:
-            host = server_addr[0]
-        else:
-            raise ValueError("The server's address is not in a valid "
-                             "'host:port' format!")
-    except Exception:
-        raise ValueError("The specified server URL is invalid.")
-
-    LOG.debug("Result: With '%s' on server '%s:%s'",
-              protocol, host, port)
-
-    return protocol, host, port
-
-
-def create_product_url(protocol, host, port, endpoint):
-    return "{0}://{1}:{2}{3}".format(protocol, host, str(port), endpoint)
-
-
-def split_product_url(url):
-    """
-    Splits the given CodeChecker server's product-specific URL into its parts.
-
-    The format of a valid URL is:
-      protocol://host:port/ProductEndpoint
-    where
-      * Protocol: HTTP or HTTPS
-      * Host: The server's host name or IP address
-      * Port: The server's port number
-      * ProductEndpoint: The product's unique endpoint folder under the server.
-
-    As a shortcut, the following formats are also valid:
-      ProductEndpoint           (means: http://localhost:8001/ProductEndpoint)
-      hostname/ProductEndpoint  (means: http://hostname:8001/ProductEndpoint)
-    """
-
-    LOG.debug("Parsing product url '%s'", url)
-    protocol, url = __strip_protocol_from_url(url)
-
-    # A valid product_url looks like this: 'http://localhost:8001/Product'.
-    protocol, port = expand_whole_protocol_and_port(protocol, None)
-    host, product_name = 'localhost', 'Default'
-    try:
-        parts = url.split("/")
-
-        if len(parts) == 1:
-            # If only one word is given in the URL, consider it as product
-            # name, but then it must appear to be a valid product name.
-
-            # "Only one word" URLs can be just simple host names too:
-            # http://codechecker.example.com:1234 should NOT be understood as
-            # the "codechecker.example.com:1234" product on "localhost:8001".
-            product_name = parts[0]
-            if product_name[0].isdigit() or '.' in product_name \
-                    or ':' in product_name:
-                raise ValueError("The given product URL is invalid. Please "
-                                 "specify a full product URL.")
-        elif len(parts) == 2:
-            # URL is at least something/product-name.
-            product_name = parts[1]
-
-            # Something is either a hostname, or a host:port.
-            server_addr = parts[0].split(":")
-            if len(server_addr) == 2:
-                host, port = server_addr[0], int(server_addr[1])
-            elif len(server_addr) == 1:
-                # We consider "localhost/product" as "localhost:8001/product".
-                host = server_addr[0]
-            else:
-                raise ValueError("The server's address is not in a valid "
-                                 "'host:port' format!")
-        else:
-            raise ValueError("Product URL can not contain extra '/' chars.")
-    except Exception:
-        raise ValueError("The specified product URL is invalid.")
-
-    LOG.debug("Result: With '%s' on server '%s:%s', product '%s'",
-              protocol, host, port, product_name)
-
-    return protocol, host, port, product_name
-
-
 def arg_match(options, args):
     """Checks and selects the option string specified in 'options'
     that are present in parameter 'args'."""
@@ -475,26 +209,6 @@ def arg_match(options, args):
             continue
 
     return matched_args
-
-
-def sizeof_fmt(num, suffix='B'):
-    """
-    Pretty print storage units.
-    Source: https://stackoverflow.com/questions/1094841/
-        reusable-library-to-get-human-readable-version-of-file-size
-    """
-    for unit in ['', 'Ki', 'Mi', 'Gi', 'Ti', 'Pi', 'Ei', 'Zi']:
-        if abs(num) < 1024.0:
-            return "%3.1f%s%s" % (num, unit, suffix)
-        num /= 1024.0
-    return "%.1f%s%s" % (num, 'Yi', suffix)
-
-
-def escape_like(string, escape_char='*'):
-    """Escape the string parameter used in SQL LIKE expressions."""
-    return string.replace(escape_char, escape_char * 2) \
-                 .replace('%', escape_char + '%') \
-                 .replace('_', escape_char + '_')
 
 
 def get_line(file_name, line_no, errors='ignore'):
@@ -522,60 +236,6 @@ def get_line(file_name, line_no, errors='ignore'):
     except IOError:
         LOG.error("Failed to open file %s", file_name)
         return u''
-
-
-class TemporaryDirectory(object):
-    def __init__(self, suffix='', prefix='tmp', tmp_dir=None):
-        self._closed = False
-        self.name = tempfile.mkdtemp(suffix, prefix, tmp_dir)
-
-    def __enter__(self):
-        return self.name
-
-    def __cleanup(self):
-        if self.name and not self._closed:
-            try:
-                shutil.rmtree(self.name)
-            except (TypeError, AttributeError) as ex:
-                print("ERROR: {0} while cleaning up {1}".format(ex, self.name))
-                return
-            self._closed = True
-
-    def __exit__(self, *args):
-        self.__cleanup()
-
-    def __del__(self, *args):
-        self.__cleanup()
-
-
-def get_user_input(msg):
-    """
-    Get the user input.
-
-    :returns: True/False based on the asnwer from the user
-    """
-    return raw_input(msg).lower() in ['y', 'yes']
-
-
-def check_file_owner_rw(file_to_check):
-    """
-    Check the file permissions.
-    Return:
-        True if only the owner can read or write the file.
-        False if other users or groups can read or write the file.
-    """
-    mode = os.stat(file_to_check)[stat.ST_MODE]
-    if mode & stat.S_IRGRP \
-            or mode & stat.S_IWGRP \
-            or mode & stat.S_IROTH \
-            or mode & stat.S_IWOTH:
-        LOG.warning("'%s' is readable by users other than you! "
-                    "This poses a risk of leaking sensitive "
-                    "information, such as passwords, session tokens, etc.!\n"
-                    "Please 'chmod 0600 %s' so only you can access the file.",
-                    file_to_check, file_to_check)
-        return False
-    return True
 
 
 def load_json_or_empty(path, default=None, kind=None, lock=False):
@@ -616,16 +276,6 @@ def load_json_or_empty(path, default=None, kind=None, lock=False):
     return ret
 
 
-def get_file_content_hash(file_path):
-    """
-    Return the file content hash for a file.
-    """
-    with open(file_path) as content:
-        hasher = hashlib.sha256()
-        hasher.update(content.read())
-        return hasher.hexdigest()
-
-
 def get_last_mod_time(file_path):
     """
     Return the last modification time of a file.
@@ -663,35 +313,3 @@ def trim_path_prefixes(path, prefixes):
         return path
 
     return path[len(longest_matching_prefix):]
-
-
-def generate_session_token():
-    """
-    Returns a random session token.
-    """
-    return uuid.UUID(bytes=os.urandom(16)).hex
-
-
-def slugify(text):
-    """
-    Removes and replaces special characters in a given text.
-    """
-    # Removes non-alpha characters.
-    norm_text = re.sub(r'[^\w\s\-/]', '', text)
-
-    # Converts spaces and slashes to underscores.
-    norm_text = re.sub(r'([\s]+|[/]+)', '_', norm_text)
-
-    return norm_text
-
-
-def replace_env_var(cfg_file):
-    def replacer(matchobj):
-        env_var = matchobj.group(1)
-        if env_var not in os.environ:
-            LOG.error('%s environment variable not set in %s', env_var,
-                      cfg_file)
-            return ''
-        return os.environ[env_var]
-
-    return replacer
