+++ web/client/codechecker_client/authentication_helper.py
@@ -15,10 +15,10 @@ from thrift.protocol import TJSONProtocol
 
 from Authentication_v6 import codeCheckerAuthentication
 
-from codechecker_common import util
 from codechecker_common.logger import get_logger
 
 from .credential_manager import SESSION_COOKIE_NAME
+from .product import create_product_url
 from .thrift_call import ThriftClientCall
 
 LOG = get_logger('system')
@@ -29,7 +29,7 @@ class ThriftAuthHelper(object):
                  session_token=None):
         self.__host = host
         self.__port = port
-        url = util.create_product_url(protocol, host, port, uri)
+        url = create_product_url(protocol, host, port, uri)
         self.transport = THttpClient.THttpClient(url)
         self.protocol = TJSONProtocol.TJSONProtocol(self.transport)
         self.client = codeCheckerAuthentication.Client(self.protocol)
