+++ web/client/codechecker_client/credential_manager.py
@@ -18,9 +18,9 @@ import stat
 import portalocker
 
 from codechecker_common.logger import get_logger
-from codechecker_common.util import check_file_owner_rw
 from codechecker_common.util import load_json_or_empty
 
+from codechecker_web.shared.env import check_file_owner_rw
 from codechecker_web.shared.version import SESSION_COOKIE_NAME as _SCN
 
 LOG = get_logger('system')
