+++ web/server/codechecker_server/api/authentication.py
@@ -17,12 +17,14 @@ from Authentication_v6.ttypes import *
 
 from codechecker_common.logger import get_logger
 from codechecker_common.profiler import timeit
-from codechecker_common.util import generate_session_token, DBSession
 
 from ..database.config_db_model import Session
 from ..permissions import handler_from_scope_params as make_handler
 from ..permissions import require_manager, require_permission
 from ..server import permissions
+from ..session_manager import generate_session_token
+
+from .db import DBSession
 
 LOG = get_logger('server')
